UPDATE ::: Removing opensearch python custom code UPDATE ::: Getting open search to work via rest and switching to that as base OPENSEARCH type UPDATE ::: Adding file exists check for DeleteFilesFromEngineRunner UPDATE ::: Minor updates to python docs UPDATE ::: Minor fix for GetEngineUsageReactor that embeddings vector file is a csv UPDATE ::: HttpHelperUtility change for HEAD request to return the headers for success (200) response code since HEAD doesn't return any data UPDATE ::: Removing unused imports
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@51126 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: ceaffb8aadcda9c259b99f22479e60433e9d004a

+++ src/prerna/cluster/util/DeleteFilesFromEngineRunner.java
@@ -1,5 +1,7 @@
 package prerna.cluster.util;
 
+import java.io.File;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -24,7 +26,10 @@ public class DeleteFilesFromEngineRunner implements Runnable {
 	public void run() {
 		for (int i = 0; i < this.FILE_PATHS.length; i++) {
 			try {
-				ClusterUtil.deleteEngineCloudFile(ENGINE_ID, ENGINE_TYPE, FILE_PATHS[i]);
+				File f = new File(FILE_PATHS[i]);
+				if(f.exists()) {
+					ClusterUtil.deleteEngineCloudFile(ENGINE_ID, ENGINE_TYPE, FILE_PATHS[i]);
+				}
 			} catch (Exception e) {
 				classLogger.error(Constants.STACKTRACE, e);
 			}
